requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
requests
boto3
psycopg2-binary
python-dotenv
